ESM3 sequence parity: unblock the fixture, then fix what it caught (ferritin-100.29, ferritin-100.31)
The fixture generator could not produce esm3_parity: the checkpoint is BFloat16
but ESM3.forward synthesises float32 plddt values and identity frames when given
no structure, and each meets a BFloat16 weight. Rather than monkeypatch the two
sites, the generator upcasts the model to float32. That fixes the class rather
than the instances, and it is the reference we actually want - the Rust side
loads these same weights at F32, so a BFloat16 reference would fold BFloat16
rounding into the thing parity is measured against. bf16 -> f32 is lossless.

The fixture then immediately failed, at cosine 0.80-0.89 per residue with the
BOS/EOS rows at roughly 15x the reference norm. Two independent defects, both in
the input-embedding stage:

`EncodeInputs::forward` skipped absent tracks, treating "no data" as "no
contribution". Upstream fills every missing track with its pad/mask token and
embeds it, so a sequence-only forward sums four non-zero contributions, not one.
The two plddt projections dominate the input embedding (norm ~14.4 against
2-4 for sequence), and the structure track is position-dependent because
sequence BOS/EOS map to STRUCTURE_BOS/EOS rather than mask - which is why those
two rows were the worst.

Both plddt projections then turned out to load `linear_no_bias` against an
upstream `nn.Linear(16, d_model)` that has one, dropping two tensors that are in
the checkpoint. Same defect as ferritin-100.27's structure-encoder FFN biases. It
hid because the first bug meant these projections were never called at all.

Everything downstream was already exact: monkeypatching the SDK to skip the same
tracks reproduced ferritin's old output at cosine 1.000000 and matching norms to
6+ digits at all 11 positions, so the tokenizer, transformer stack, geometric
attention and output heads needed no changes. With both fixed, parity is
0.99982-0.99993 per position.

`encode_inputs` also now casts the rbf-encoded plddt to the projection's weight
dtype, so a half-precision ESM3 does not hit the same mismatch that breaks the
SDK on its own BF16 checkpoint.

esm3-sm-open-v1 moves to ParityStatus::Verified.

+++ scripts/generate_esm3_fixtures.py
@@ -70,8 +70,23 @@ def get_embeddings(sequence: str) -> torch.Tensor:
     from esm.models.esm3 import ESM3
     from esm.sdk.api import ESMProtein, SamplingConfig
 
+    # Upcast to float32 before running (ferritin-100.29).
+    #
+    # The checkpoint is BFloat16, but ESM3.forward synthesises float32
+    # auxiliary tensors when the caller supplies no structure — the plddt
+    # values feeding `EncodeInputs.plddt_projection`, and the identity frames
+    # feeding geometric attention. In esm 3.4.0 / torch 2.11 each meets a
+    # BFloat16 weight and raises "expected m1 and m2 to have the same dtype".
+    # Both are upstream SDK bugs, not ferritin's.
+    #
+    # Upcasting fixes the class rather than the two instances, and it is also
+    # the reference we actually want: the Rust side loads these same weights at
+    # F32, so a BFloat16 reference would fold BFloat16 rounding into the thing
+    # parity is measured against. bf16 -> f32 is lossless, so no checkpoint
+    # information is invented by doing this.
     client = ESM3.from_pretrained("esm3_sm_open_v1")
     client.eval()
+    client = client.to(torch.float32)
 
     protein = ESMProtein(sequence=sequence)
     tensor = client.encode(protein)
@@ -161,11 +176,8 @@ def main():
     output_dir.mkdir(parents=True, exist_ok=True)
 
     # The two fixtures are generated independently: they exercise different
-    # checkpoints and different SDK entry points, and the sequence path is
-    # currently broken by an esm/torch dtype mismatch inside the SDK itself
-    # (`plddt_projection` receives Float against BFloat16 weights). Letting
-    # that failure abort the run would also lose the structure fixture, which
-    # is unrelated to it and does work.
+    # checkpoints and different SDK entry points, so a failure in one should
+    # not cost the other. Both currently generate.
     failures = []
 
     try:
